--- a/pyproject.toml
+++ b/pyproject.toml
@@ -13,6 +13,7 @@ requires-python = ">=3.8"
 dependencies = [
   "ga4mp",
   "optuna",
+  "GitPython",
   "PyGithub",
 ]
 dynamic = ["version"]