## Motivation
<!-- Describe your motivation why you will submit this PR. This is useful for reviewers to understand the context of PR. -->

Currently, the `load_module` function often gets stuck due to [the rate limit for the GitHub REST API](https://docs.github.com/en/rest/using-the-rest-api/rate-limits-for-the-rest-api?apiVersion=2022-11-28).
The rate limit for unauthenticated requests is 60 requests per hour, which is very strict.

The error message looks like this:
```
Request GET /repos/optuna/optunahub-registry failed with 403: rate limit exceeded
INFO:github.GithubRetry:Request GET /repos/optuna/optunahub-registry failed with 403: rate limit exceeded
Setting next backoff to 3570.799172s
INFO:github.GithubRetry:Setting next backoff to 3570.799172s
```

## Description of the changes
<!-- Describe the changes in this PR. -->

- Use `git` command when available to avoid the rate limit.

## Benchmarking results

I benchmarked:
- How many times `samplers/auto_sampler` can be loaded within one minute
- The time it takes to load once

### Setups

The benchmark code I used here is as follows:
```py
!pip install optuna
# or
# !pip install git+https://github.com/kAIto47802/optunahub@reduce-api-call-limit

!pip install -q cmaes


import optuna
import optunahub
import time

start = time.time()
for i in range(500):
    mod = optunahub.load_module("samplers/auto_sampler", force_reload=True)
    print(i)
    end = time.time()
    print("end - start:", end - start)
    if end - start > 60:
        break

print("end - start:", end - start)
print("Number of iterations:", i)
```

The benchmarks were run five times for both the original and the new implementation.

I used a different Google Colaboratory notebook for each benchmarking run to ensure a different IP address was used every time. (Also, I confirmed the attached VM's IP address is certainly different by using the command `curl ifconfig.me`.)


### Results

The average number of times `load_module` was successfully executed without hitting the API rate limit is as follows:

|Original| This PR|
|---|---|
|5.2| - |

With the new implementation, running continuous downloads for one minute does not hit a rate limit across all five runs. (Therefore, it is indicated with a dash (–) in the table.)


The average time for downloading once is as follows:

|Original|This PR|
|---|---|
|4.50 ± 0.45 sec|0.49 ± 0.12 sec|

$m \pm \sigma$ indicates the average and the standard error over the five runs, respectively.