@nabenabe0928 2025-05-14
I confirmed that the change worked nicely!

```bash
$ pip install git+https://github.com/kAIto47802/optunahub.git@reduce-api-call-limit
$ python -c "import optunahub; optunahub.load_module("samplers/ctpe")"
$ ls .cache/optunahub/github.com/optuna/optunahub-registry/main/package/samplers/
ctpe
```